requirements.txt
#requirements for this experiment
otree==5.10.3
psycopg2>=2.8.4
pandas~=1.5.2
numpy~=1.24.1
requests~=2.31.0
Pillow~=9.5.0